Create Tables database and modify table status type
In this commit, a new SQL file is created to initialize Tables database with appropriate fields and status enums. The existing enum, TableStatusType, is modified to match the new status present in the Tables database. Some changes in mysql configuration are also made in compose.yaml and application.yaml files, alongside the addition of dependencies for Flyway in the pom.xml file.