Refactor table service and add new use cases
This commit refactored the Table service by renaming and restructuring some classes. The 'TablesPort' interface was renamed to 'TableDatasourcePort' and 'TableConditionPort' was renamed to 'TableConditionDatasourcePort'. Also, the class 'GetTablesUseCase' was updated to use the new 'TableDatasourcePort'. Two new use cases were added - 'GetNotAvailableTimeForReservationUseCase' and 'ReserveTableUseCase'. The 'ReserveTableProcessor' was also moved to the 'processor' package and updated to use 'TableConditionDatasourcePort'.